+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 1.7.2  *(2016-06-26)*
 --------------------
 * When enabling the Word of the Day, create a Wotd notification immediately.
+* Added a feature to look up a random word in the dictionary.
 
 1.7.1  *(2016-06-26)*
 --------------------